pyproject.toml
[project]
name = "velib-data-ingestion"
version = "0.1.0"
description = "Velib data ingestion and processing pipeline"
requires-python = ">=3.10"
authors = [
    { name = "Allister K.", email = "contact@allisterkohn.com" }
]

dependencies = [
    "boto3",
    "geopandas",
    "pandas",
    "polars",
    "python-dotenv",
    "requests",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py310"
line-length = 120
exclude = ["*.ipynb", "__pycache__", ".venv"]
cache-dir = "~/.cache/ruff"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501", "E731"]

[tool.ruff.lint.isort]
known-first-party = ["velib_data_ingestion"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
